Move unit tests to a directory structure that follows the classes under test
